remove old code, add tests
